requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
drf-yasg==1.21.7
setuptools==69.0.2
django-allauth==0.58.2
django-redis==5.4.0
django-cors-headers==4.3.1
django-filter==23.5
django-debug-toolbar==4.2.0
sentry-sdk==1.39.1
django-storages==1.14.2
Pillow==10.1.0
uritemplate==4.1.1
packaging==23.2
coreapi==2.3.3
PyYAML==6.0.1
